plugins/about: fix bracket mistake in DevMode page
The parameters of `Qt.resolvedUrl()` isn't supposed to include
`{securityPrivacy...` thing. So, move the bracket into the right place.
While we're at it, also format the code so that it looks better and is
easier to spot a similar kind of mistake.

I can't believe I let this slip in. Maybe I should look for a new pair
of glasses. 🤦

Fixes: 02667dfa ("Pass instance of LomiriSecurityPrivacyPanel to
LockSecurity page")