plugins/about: fix enabling/disabling developer mode while unplugged
Because `mode_request` will return a valid value when being plugged
only, the "_adb while retaining mtp or rndis" logic breaks down when the
phone is not plugged to a computer. Thus, use the `get_config` call, the
same one as `getDeveloperMode()`, instead.

And while we're at it, improve robustness of the code by handling the
"unknown mode" correctly.

Fixes: https://gitlab.com/ubports/development/core/lomiri-system-settings/-/issues/344