plugins/hotspot: remove error reporting from hotspot setup dialog
I've added the reporting in lomiri-indicator-network as a notification
which appears every time hotspot cannot be activated [1]. So, having one
in the hotspot setup dialog too is now redundant. So remove it.

This is prompted by a request from @sunweaver to avoid string
duplication with [1]. Unfortunately, the same strings are also presented
in wi-fi plugin, so this doesn't achieve that.

Fixes: https://gitlab.com/ubports/development/core/lomiri-system-settings/-/issues/349

[1] https://gitlab.com/ubports/development/core/lomiri-indicator-network/-/merge_requests/115