plugins/wifi: fix weird behavior with the peer picker for cert
When the peer picker was shown, it was shifted up by the keyboard, which
was still opened because the focus was still with at one of the text
box.

So, moves focus to the peer picker when it's opened. User has to
interact with it next, so it makes sense. That still makes the peer
picker shifting back while the keyboard hides, so also make it so that
the peer picker isn't shifted by the keyboard.

Fixes: https://gitlab.com/ubports/development/core/lomiri-system-settings/-/issues/206